Reorganise for more flexible distribution
(I did not update the Russian text in README.md, sorry!)

The configuration is now installed using `git --config`, so it will work if
the file format or location changes in future. Also, the per-repository
configuration can now be accessed from anywhere (as a simple command).

Use odt2txt for better conversion to text. It’s available in many distros
already, but give the source URL too.